requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
rembg>=2.0.67
pillow==10.4.0
onnxruntime>=1.18.0; platform_system != "Darwin" or platform_machine != "arm64"
onnxruntime-silicon>=1.18.0; platform_system == "Darwin" and platform_machine == "arm64"
python-multipart>=0.0.9
opencv-python-headless>=4.9.0.80

